return dtype is fp32 in most models

+++ src/transformers/models/ernie4_5_vl_moe/modeling_ernie4_5_vl_moe.py
@@ -897,7 +897,7 @@ class Ernie4_5_VLMoeVisionRotaryEmbedding(nn.Module):
 
         cos = self.recomposition_frequencies(cos)
         sin = self.recomposition_frequencies(sin)
-        return cos.to(dtype=x.dtype), sin.to(dtype=x.dtype)
+        return cos, sin
 
     def recomposition_frequencies(self, freq):
         """
