+++ src/transformers/models/exaone4_5/modeling_exaone4_5.py
@@ -130,7 +130,7 @@ class Exaone4_5_VisionRotaryEmbedding(nn.Module):
 
         cos = self.recomposition_frequencies(cos)
         sin = self.recomposition_frequencies(sin)
-        return cos.to(dtype=x.dtype), sin.to(dtype=x.dtype)
+        return cos, sin
 
     def recomposition_frequencies(self, freq):
         """
